Refactor NotesRender.tsx into separate components and create custom hooks for note management, drag-and-drop functionality, and dialog handling
